# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zero-Click CRM API
  version: 1.0.0
  description: >-
    Documentación de la API para el backend Zero-Click CRM. Incluye ingestión multicanal,
    procesamiento de audio/video, consultas CRM, chat asistido por IA, alertas, notas,
    knowledge base y endpoints administrativos.
  contact:
    name: Equipo Relay
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo local
  - url: https://your-deployment-url.example
    description: Servidor de producción (actualiza esta URL al desplegar)
...
paths:
...
  /api/alerts:
    get:
      summary: Listar alertas
      tags: [Alerts]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: open
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/AlertSeverity'
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: contactId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Alertas paginadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsListResponse'
  /api/alerts/{id}/resolve:
    post:
      summary: Marcar una alerta como resuelta
      tags: [Alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alerta resuelta
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/seed/usuarios:
    post:
      summary: Seed de contactos/usuarios (usa mock JSON o Python generator)
      tags: [Admin]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                generate:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Seed ejecutado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  result:
                    type: object
                    properties:
                      count:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/seed/completo:
    post:
      summary: Seed completo (companies, contacts, work items, interacciones, fresh data)
      tags: [Admin]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                generate:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Seed ejecutado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  result:
                    type: object
                    properties:
                      companies:
                        type: integer
                      contacts:
                        type: integer
                      workItems:
                        type: integer
                      interactions:
                        type: integer
                      freshData:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/analyze/trigger:
    post:
      summary: Disparar análisis manual sobre tablas soportadas
      tags: [Admin]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [interactions, work_items, contacts, fresh_data]
                  default: interactions
                id:
                  type: string
                limit:
                  type: integer
                  default: 10
      responses:
        '200':
          description: Número de registros analizados
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  analyzed:
                    type: integer
...
  /api/admin/ai/reindex:
    post:
      summary: Reindexar embeddings/contextos para IA
      tags: [Admin]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [interactions, work_items, fresh_data, all]
                  default: all
                limit:
                  type: integer
                  default: 100
                companyId:
                  type: string
      responses:
        '200':
          description: Reindex ejecutado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  type:
                    type: string
                  result:
                    type: object
components:
  schemas:
...
    Sentiment:
      type: string
      enum: [positive, neutral, negative]
...
    AlertSeverity:
      type: string
      enum: [low, medium, high, critical]
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        domain:
          type: string
          nullable: true
...
    ContactBasic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        sentiment:
          $ref: '#/components/schemas/Sentiment'
          nullable: true
...
    Alert:
      type: object
      properties:
        id:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        status:
          type: string
        message:
          type: string
        data:
          type: object
        company:
          $ref: '#/components/schemas/Company'
        contact:
          $ref: '#/components/schemas/ContactBasic'
...
    AlertsListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        count:
          type: integer
...
  parameters:
    CompanyIdQuery:
      name: companyId
      in: query
      schema:
        type: string
      description: Filtra por UUID de la compañía
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    OffsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
...